+++ php/src/Application/Command.php
@@ -23,7 +23,7 @@ abstract class Command implements CommandInterface
      *
      * @var int
      */
-    const DATABASE_VERSION = 14;
+    const DATABASE_VERSION = 15;
 
     /**
      * @var IndexDatabase