+++ php/tests/Application/Command/GlobalConstantsTest.php
@@ -18,7 +18,9 @@ class GlobalConstantsTest extends IndexedTest
         $output = $command->getGlobalConstants();
 
         $this->assertThat($output, $this->arrayHasKey('FIRST_CONSTANT'));
+        $this->assertEquals($output['FIRST_CONSTANT']['fqsen'], 'A\FIRST_CONSTANT');
         $this->assertThat($output, $this->arrayHasKey('SECOND_CONSTANT'));
+        $this->assertEquals($output['SECOND_CONSTANT']['fqsen'], 'A\SECOND_CONSTANT');
         $this->assertThat($output, $this->logicalNot($this->arrayHasKey('SHOULD_NOT_SHOW_UP')));
     }
 }